# lib/api-zod/src/generated/types/createEmployerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SSS Ormoc Digital R-1/R-1A Repository API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEmployerBody {
  employerId: string;
  name: string;
  address: string;
  jurisdiction: string;
}