# lib/api-zod/src/generated/types/jurisdictionList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SSS Ormoc Digital R-1/R-1A Repository API
 * OpenAPI spec version: 0.1.0
 */
import type { JurisdictionCategory } from "./jurisdictionCategory";

export interface JurisdictionList {
  categories: JurisdictionCategory[];
}