# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SSS Ormoc Digital R-1/R-1A Repository API
 * OpenAPI spec version: 0.1.0
 */

export interface SuccessResponse {
  message: string;
}